+++ pkg/query/columnquery_test.go
@@ -1230,8 +1230,6 @@ func PprofToSymbolizedProfile(meta profile.Meta, prof *pprofprofile.Profile, ind
 		labelNames = append(labelNames, l)
 	}
 
-	durationSeconds := float64(meta.Duration) / float64(time.Second.Nanoseconds())
-
 	w := profile.NewWriter(memory.DefaultAllocator, labelNames)
 	defer w.RecordBuilder.Release()
 	for i := range prof.Sample {
@@ -1240,14 +1238,7 @@ func PprofToSymbolizedProfile(meta profile.Meta, prof *pprofprofile.Profile, ind
 		}
 
 		w.Value.Append(prof.Sample[i].Value[index])
-		if meta.Duration > 0 {
-			w.ValuePerSecond.Append(float64(prof.Sample[i].Value[index]) / durationSeconds)
-		} else {
-			w.ValuePerSecond.AppendNull()
-		}
-
 		w.Diff.Append(0)
-		w.DiffPerSecond.Append(0)
 
 		for labelName, labelBuilder := range w.LabelBuildersMap {
 			if prof.Sample[i].Label == nil {
@@ -1326,9 +1317,7 @@ func OldProfileToArrowProfile(p profile.OldProfile) (profile.Profile, error) {
 	defer w.RecordBuilder.Release()
 	for i := range p.Samples {
 		w.Value.Append(p.Samples[i].Value)
-		w.ValuePerSecond.Append(float64(p.Samples[i].Value))
 		w.Diff.Append(p.Samples[i].DiffValue)
-		w.DiffPerSecond.Append(float64(p.Samples[i].DiffValue))
 
 		for labelName, labelBuilder := range w.LabelBuildersMap {
 			if p.Samples[i].Label == nil {
@@ -1440,9 +1429,7 @@ func TestFilterData(t *testing.T) {
 	w.FunctionFilename.Append([]byte("test"))
 	w.FunctionStartLine.Append(1)
 	w.Value.Append(1)
-	w.ValuePerSecond.Append(1)
 	w.Diff.Append(0)
-	w.DiffPerSecond.Append(0)
 
 	frameFilter := map[string]struct{}{"test": {}}
 	originalRecord := w.RecordBuilder.NewRecord()
@@ -1488,9 +1475,7 @@ func TestFilterUnsymbolized(t *testing.T) {
 	w.MappingBuildID.Append([]byte("test"))
 	w.Lines.Append(false)
 	w.Value.Append(1)
-	w.ValuePerSecond.Append(1)
 	w.Diff.Append(0)
-	w.DiffPerSecond.Append(0)
 
 	originalRecord := w.RecordBuilder.NewRecord()
 	recs, _, err := FilterProfileData(
@@ -1570,9 +1555,7 @@ func TestFilterDataWithPath(t *testing.T) {
 	w.FunctionFilename.Append([]byte("test.py"))
 	w.FunctionStartLine.Append(0)
 	w.Value.Append(1)
-	w.ValuePerSecond.Append(1)
 	w.Diff.Append(0)
-	w.DiffPerSecond.Append(0)
 
 	frameFilter := map[string]struct{}{"libpython3.11.so.1.0": {}, "interpreter": {}}
 	originalRecord := w.RecordBuilder.NewRecord()
@@ -1654,9 +1637,7 @@ func TestFilterDataFrameFilter(t *testing.T) {
 	w.FunctionFilename.Append([]byte("test.py"))
 	w.FunctionStartLine.Append(0)
 	w.Value.Append(1)
-	w.ValuePerSecond.Append(1)
 	w.Diff.Append(0)
-	w.DiffPerSecond.Append(0)
 
 	frameFilter := map[string]struct{}{"interpreter": {}}
 	originalRecord := w.RecordBuilder.NewRecord()
@@ -1738,9 +1719,7 @@ func BenchmarkFilterData(t *testing.B) {
 		w.FunctionFilename.Append([]byte("test"))
 		w.FunctionStartLine.Append(1)
 		w.Value.Append(1)
-		w.ValuePerSecond.Append(1)
 		w.Diff.Append(0)
-		w.DiffPerSecond.Append(0)
 	}
 
 	originalRecord := w.RecordBuilder.NewRecord()