+++ pkg/query/flamegraph_arrow.go
@@ -53,14 +53,11 @@ const (
 	FlamegraphFieldFunctionSystemName = "function_system_name"
 	FlamegraphFieldFunctionFileName   = "function_file_name"
 
-	FlamegraphFieldLabels              = "labels"
-	FlamegraphFieldChildren            = "children"
-	FlamegraphFieldCumulative          = "cumulative"
-	FlamegraphFieldCumulativePerSecond = "cumulative_per_second"
-	FlamegraphFieldFlat                = "flat"
-	FlamegraphFieldFlatPerSecond       = "flat_per_second"
-	FlamegraphFieldDiff                = "diff"
-	FlamegraphFieldDiffPerSecond       = "diff_per_second"
+	FlamegraphFieldLabels     = "labels"
+	FlamegraphFieldChildren   = "children"
+	FlamegraphFieldCumulative = "cumulative"
+	FlamegraphFieldFlat       = "flat"
+	FlamegraphFieldDiff       = "diff"
 )
 
 func GenerateFlamegraphArrow(
@@ -127,9 +124,7 @@ func generateFlamegraphArrowRecord(ctx context.Context, mem memory.Allocator, tr
 	labelHasher := xxh3.New()
 	for _, r := range profileReader.RecordReaders {
 		fb.cumulative += math.Int64.Sum(r.Value)
-		fb.cumulativePerSecond += math.Float64.Sum(r.ValuePerSecond)
 		fb.diff += math.Int64.Sum(r.Diff)
-		fb.diffPerSecond += math.Float64.Sum(r.DiffPerSecond)
 
 		if err := fb.ensureLabelColumns(r.LabelFields); err != nil {
 			return nil, 0, 0, 0, fmt.Errorf("ensure label columns: %w", err)
@@ -350,16 +345,10 @@ func generateFlamegraphArrowRecord(ctx context.Context, mem memory.Allocator, tr
 		fb.trimmedFunctionStartLine.AppendNull()
 		fb.trimmedCumulative = array.NewUint8Builder(fb.pool)
 		fb.trimmedCumulative.AppendNull()
-		fb.trimmedCumulativePerSecond = builder.NewOptFloat64Builder(arrow.PrimitiveTypes.Float64)
-		fb.trimmedCumulativePerSecond.AppendNull()
 		fb.trimmedFlat = array.NewUint8Builder(fb.pool)
 		fb.trimmedFlat.AppendNull()
-		fb.trimmedFlatPerSecond = builder.NewOptFloat64Builder(arrow.PrimitiveTypes.Float64)
-		fb.trimmedFlatPerSecond.AppendNull()
 		fb.trimmedDiff = array.NewUint8Builder(fb.pool)
 		fb.trimmedDiff.AppendNull()
-		fb.trimmedDiffPerSecond = builder.NewOptFloat64Builder(arrow.PrimitiveTypes.Float64)
-		fb.trimmedDiffPerSecond.AppendNull()
 	}
 
 	_, spanNewRecord := tracer.Start(ctx, "NewRecord")
@@ -697,13 +686,10 @@ func (fb *flamegraphBuilder) mergeUnsymbolizedRows(
 		}
 
 		value := r.Value.Value(sampleIndex)
-		valuePerSecond := r.ValuePerSecond.Value(sampleIndex)
 
 		fb.builderCumulative.Add(cr, value)
-		fb.builderCumulativePerSecond.Add(cr, valuePerSecond)
 		if leaf {
 			fb.builderFlat.Add(cr, value)
-			fb.builderFlatPerSecond.Add(cr, valuePerSecond)
 		}
 
 		fb.parent.Set(cr)
@@ -771,12 +757,8 @@ type flamegraphBuilder struct {
 
 	// This keeps track of the total cumulative value so that we can set the first row's cumulative value at the end.
 	cumulative int64
-	// This keeps track of the total cumulative per second value so that we can set the first row's cumulative value at the end.
-	cumulativePerSecond float64
 	// This keeps track of the total diff values so that we can set the first row's diff value at the end.
 	diff int64
-	// This keeps track of the total diff per second value so that we can set the first row's diff value at the end.
-	diffPerSecond float64
 	// This keeps track of the max height of the flame graph.
 	maxHeight int32
 	// trimmed keeps track of the values that were trimmed from the flame graph.
@@ -819,11 +801,8 @@ type flamegraphBuilder struct {
 	builderChildren                      *builder.ListBuilder
 	builderChildrenValues                *array.Uint32Builder
 	builderCumulative                    *builder.OptInt64Builder
-	builderCumulativePerSecond           *builder.OptFloat64Builder
 	builderFlat                          *builder.OptInt64Builder
-	builderFlatPerSecond                 *builder.OptFloat64Builder
 	builderDiff                          *builder.OptInt64Builder
-	builderDiffPerSecond                 *builder.OptFloat64Builder
 
 	// Only at the last step when preparing the new record these are populated.
 	// They are also used to create compacted dictionaries and after that replaced by them.
@@ -843,14 +822,11 @@ type flamegraphBuilder struct {
 
 	labelNameIndex map[string]int
 
-	trimmedLocationLine        array.Builder
-	trimmedFunctionStartLine   array.Builder
-	trimmedCumulative          array.Builder
-	trimmedCumulativePerSecond *builder.OptFloat64Builder
-	trimmedFlat                array.Builder
-	trimmedFlatPerSecond       *builder.OptFloat64Builder
-	trimmedDiff                array.Builder
-	trimmedDiffPerSecond       *builder.OptFloat64Builder
+	trimmedLocationLine      array.Builder
+	trimmedFunctionStartLine array.Builder
+	trimmedCumulative        array.Builder
+	trimmedFlat              array.Builder
+	trimmedDiff              array.Builder
 }
 
 type aggregationConfig struct {
@@ -903,14 +879,11 @@ func newFlamegraphBuilder(
 		builderFunctionFilenameIndices:       array.NewInt32Builder(pool),
 		builderFunctionFilenameDictUnifier:   array.NewBinaryDictionaryUnifier(pool),
 
-		builderChildren:            builderChildren,
-		builderChildrenValues:      builderChildren.ValueBuilder().(*array.Uint32Builder),
-		builderCumulative:          builder.NewOptInt64Builder(arrow.PrimitiveTypes.Int64),
-		builderCumulativePerSecond: builder.NewOptFloat64Builder(arrow.PrimitiveTypes.Float64),
-		builderFlat:                builder.NewOptInt64Builder(arrow.PrimitiveTypes.Int64),
-		builderFlatPerSecond:       builder.NewOptFloat64Builder(arrow.PrimitiveTypes.Float64),
-		builderDiff:                builder.NewOptInt64Builder(arrow.PrimitiveTypes.Int64),
-		builderDiffPerSecond:       builder.NewOptFloat64Builder(arrow.PrimitiveTypes.Float64),
+		builderChildren:       builderChildren,
+		builderChildrenValues: builderChildren.ValueBuilder().(*array.Uint32Builder),
+		builderCumulative:     builder.NewOptInt64Builder(arrow.PrimitiveTypes.Int64),
+		builderFlat:           builder.NewOptInt64Builder(arrow.PrimitiveTypes.Int64),
+		builderDiff:           builder.NewOptInt64Builder(arrow.PrimitiveTypes.Int64),
 	}
 
 	for _, f := range aggregateFields {
@@ -950,12 +923,9 @@ func newFlamegraphBuilder(
 
 	// The cumulative values is calculated and at the end set to the correct value.
 	fb.builderCumulative.Append(0)
-	fb.builderCumulativePerSecond.Append(0)
 	fb.builderDiff.Append(0)
-	fb.builderDiffPerSecond.Append(0)
 	// the root will never have a flat value
 	fb.builderFlat.Append(0)
-	fb.builderFlatPerSecond.Append(0)
 
 	return fb, nil
 }
@@ -973,9 +943,7 @@ func (fb *flamegraphBuilder) prepareNewRecord() error {
 	// Now we set/overwrite the cumulative value for the root row (which is always the 0 row in our flame graphs).
 	// We don't care about a global flat value, therefore it's omitted here.
 	fb.builderCumulative.Set(0, fb.cumulative)
-	fb.builderCumulativePerSecond.Set(0, fb.cumulativePerSecond)
 	fb.builderDiff.Set(0, fb.diff)
-	fb.builderDiffPerSecond.Set(0, fb.diffPerSecond)
 
 	// We want to unify the dictionaries after having created the flame graph now.
 	// They are going to be trimmed and compacted in the next step.
@@ -1065,7 +1033,7 @@ func (fb *flamegraphBuilder) prepareNewRecord() error {
 // It adds the children to the children column and the labels intersection to the labels column.
 // Finally, it assembles all columns from the builders into an arrow record.
 func (fb *flamegraphBuilder) NewRecord() (arrow.Record, error) {
-	const numCols = 17
+	const numCols = 14
 
 	cleanupArrs := make([]arrow.Array, 0, numCols+1+(2*len(fb.builderLabelFields)))
 	defer func() {
@@ -1108,11 +1076,8 @@ func (fb *flamegraphBuilder) NewRecord() (arrow.Record, error) {
 		// Values
 		{Name: FlamegraphFieldChildren, Type: arrow.ListOf(arrow.PrimitiveTypes.Uint32)},
 		{Name: FlamegraphFieldCumulative, Type: fb.trimmedCumulative.Type()},
-		{Name: FlamegraphFieldCumulativePerSecond, Type: arrow.PrimitiveTypes.Float64},
 		{Name: FlamegraphFieldFlat, Type: fb.trimmedFlat.Type()},
-		{Name: FlamegraphFieldFlatPerSecond, Type: arrow.PrimitiveTypes.Float64},
 		{Name: FlamegraphFieldDiff, Type: fb.trimmedDiff.Type()},
-		{Name: FlamegraphFieldDiffPerSecond, Type: arrow.PrimitiveTypes.Float64},
 	}
 
 	arrays := make([]arrow.Array, numCols+len(fb.labels))
@@ -1135,16 +1100,10 @@ func (fb *flamegraphBuilder) NewRecord() (arrow.Record, error) {
 	cleanupArrs = append(cleanupArrs, arrays[10])
 	arrays[11] = fb.trimmedCumulative.NewArray()
 	cleanupArrs = append(cleanupArrs, arrays[11])
-	arrays[12] = fb.trimmedCumulativePerSecond.NewArray()
+	arrays[12] = fb.trimmedFlat.NewArray()
 	cleanupArrs = append(cleanupArrs, arrays[12])
-	arrays[13] = fb.trimmedFlat.NewArray()
+	arrays[13] = fb.trimmedDiff.NewArray()
 	cleanupArrs = append(cleanupArrs, arrays[13])
-	arrays[14] = fb.trimmedFlatPerSecond.NewArray()
-	cleanupArrs = append(cleanupArrs, arrays[14])
-	arrays[15] = fb.trimmedDiff.NewArray()
-	cleanupArrs = append(cleanupArrs, arrays[15])
-	arrays[16] = fb.trimmedDiffPerSecond.NewArray()
-	cleanupArrs = append(cleanupArrs, arrays[16])
 
 	for i, field := range fb.builderLabelFields {
 		field.Type = fb.labels[i].DataType() // overwrite for variable length uint types
@@ -1182,11 +1141,8 @@ func (fb *flamegraphBuilder) Release() {
 
 	fb.builderChildren.Release()
 	fb.builderCumulative.Release()
-	fb.builderCumulativePerSecond.Release()
 	fb.builderFlat.Release()
-	fb.builderFlatPerSecond.Release()
 	fb.builderDiff.Release()
-	fb.builderDiffPerSecond.Release()
 
 	if fb.trimmedLocationLine != nil {
 		fb.trimmedLocationLine.Release()
@@ -1202,10 +1158,6 @@ func (fb *flamegraphBuilder) Release() {
 
 	if fb.trimmedDiff != nil {
 		fb.trimmedDiff.Release()
-	}
-
-	if fb.trimmedDiffPerSecond != nil {
-		fb.trimmedDiffPerSecond.Release()
 	}
 
 	for i := range fb.builderLabelFields {
@@ -1342,29 +1294,20 @@ func (fb *flamegraphBuilder) appendRow(
 	}
 
 	value := r.Value.Value(sampleRow)
-	valuePerSecond := r.ValuePerSecond.Value(sampleRow)
 
 	fb.builderCumulative.Append(value)
-	fb.builderCumulativePerSecond.Append(valuePerSecond)
 
 	if leaf { // leaf
 		fb.builderFlat.Append(value)
-		fb.builderFlatPerSecond.Append(valuePerSecond)
 	} else {
 		// It's possible that these get merged later on and we don't want to set these to null
 		fb.builderFlat.Append(0)
-		fb.builderFlatPerSecond.Append(0)
 	}
 
 	if r.Diff.Value(sampleRow) > 0 {
 		fb.builderDiff.Append(r.Diff.Value(sampleRow))
 	} else {
 		fb.builderDiff.AppendNull()
-	}
-	if r.DiffPerSecond.Value(sampleRow) != 0 {
-		fb.builderDiffPerSecond.Append(r.DiffPerSecond.Value(sampleRow))
-	} else {
-		fb.builderDiffPerSecond.AppendNull()
 	}
 
 	return nil
@@ -1424,11 +1367,8 @@ func (fb *flamegraphBuilder) AppendLabelRow(
 
 	// Append both cumulative and diff values and overwrite them below.
 	fb.builderCumulative.Append(0)
-	fb.builderCumulativePerSecond.Append(0)
 	fb.builderDiff.Append(0)
-	fb.builderDiffPerSecond.Append(0)
 	fb.builderFlat.Append(0)
-	fb.builderFlatPerSecond.Append(0)
 	fb.addRowValues(r, row, sampleRow, leaf)
 
 	return nil
@@ -1437,16 +1377,12 @@ func (fb *flamegraphBuilder) AppendLabelRow(
 // addRowValues updates the existing row's values and potentially adding existing values on top.
 func (fb *flamegraphBuilder) addRowValues(r *profile.RecordReader, row, sampleRow int, leaf bool) {
 	value := r.Value.Value(sampleRow)
-	valuePerSecond := r.ValuePerSecond.Value(sampleRow)
 
 	fb.builderCumulative.Add(row, value)
-	fb.builderCumulativePerSecond.Add(row, valuePerSecond)
 	fb.builderDiff.Add(row, r.Diff.Value(sampleRow))
-	fb.builderDiffPerSecond.Add(row, r.DiffPerSecond.Value(sampleRow))
 
 	if leaf {
 		fb.builderFlat.Add(row, value)
-		fb.builderFlatPerSecond.Add(row, valuePerSecond)
 	}
 }
 
@@ -1520,13 +1456,10 @@ func (fb *flamegraphBuilder) trim(ctx context.Context, tracer trace.Tracer, thre
 	trimmedFunctionFilenameIndices := array.NewInt32Builder(fb.pool)
 	trimmedCumulativeType := smallestUnsignedTypeFor(largestCumulativeValue)
 	trimmedCumulative := array.NewBuilder(fb.pool, trimmedCumulativeType)
-	trimmedCumulativePerSecond := builder.NewOptFloat64Builder(arrow.PrimitiveTypes.Float64)
 	trimmedFlatType := smallestUnsignedTypeFor(largestFlatValue)
 	trimmedFlat := array.NewBuilder(fb.pool, trimmedFlatType)
-	trimmedFlatPerSecond := builder.NewOptFloat64Builder(arrow.PrimitiveTypes.Float64)
 	trimmedDiffType := smallestSignedTypeFor(smallestDiffValue, largestDiffValue)
 	trimmedDiff := array.NewBuilder(fb.pool, trimmedDiffType)
-	trimmedDiffPerSecond := builder.NewOptFloat64Builder(arrow.PrimitiveTypes.Float64)
 
 	releasers = append(releasers,
 		trimmedMappingFileIndices,
@@ -1557,11 +1490,8 @@ func (fb *flamegraphBuilder) trim(ctx context.Context, tracer trace.Tracer, thre
 	trimmedFunctionSystemNameIndices.Reserve(row)
 	trimmedFunctionFilenameIndices.Reserve(row)
 	trimmedCumulative.Reserve(row)
-	trimmedCumulativePerSecond.Reserve(row)
 	trimmedFlat.Reserve(row)
-	trimmedFlatPerSecond.Reserve(row)
 	trimmedDiff.Reserve(row)
-	trimmedDiffPerSecond.Reserve(row)
 
 	for _, l := range trimmedLabelsIndices {
 		l.Reserve(row)
@@ -1589,10 +1519,6 @@ func (fb *flamegraphBuilder) trim(ctx context.Context, tracer trace.Tracer, thre
 		for i := range fb.labels {
 			appendDictionaryIndexInt32(fb.labelsIndices[i], trimmedLabelsIndices[i], te.row)
 		}
-
-		copyOptFloat64BuilderValue(fb.builderCumulativePerSecond, trimmedCumulativePerSecond, te.row)
-		copyOptFloat64BuilderValue(fb.builderFlatPerSecond, trimmedFlatPerSecond, te.row)
-		copyOptFloat64BuilderValue(fb.builderDiffPerSecond, trimmedDiffPerSecond, te.row)
 
 		// The following two will never be null.
 		cum := fb.builderCumulative.Value(te.row)
@@ -1768,11 +1694,8 @@ func (fb *flamegraphBuilder) trim(ctx context.Context, tracer trace.Tracer, thre
 	fb.trimmedLocationLine = trimmedLocationLine
 	fb.trimmedFunctionStartLine = trimmedFunctionStartLine
 	fb.trimmedCumulative = trimmedCumulative
-	fb.trimmedCumulativePerSecond = trimmedCumulativePerSecond
 	fb.trimmedFlat = trimmedFlat
-	fb.trimmedFlatPerSecond = trimmedFlatPerSecond
 	fb.trimmedDiff = trimmedDiff
-	fb.trimmedDiffPerSecond = trimmedDiffPerSecond
 	fb.trimmedChildren = trimmedChildren
 
 	return nil
@@ -1838,14 +1761,6 @@ func copyOptBooleanBuilderValue(old, new *builder.OptBooleanBuilder, row int) {
 }
 
 func copyBoolBuilderValue(old, new *array.BooleanBuilder, row int) {
-	if old.IsNull(row) {
-		new.AppendNull()
-		return
-	}
-	new.Append(old.Value(row))
-}
-
-func copyOptFloat64BuilderValue(old, new *builder.OptFloat64Builder, row int) {
 	if old.IsNull(row) {
 		new.AppendNull()
 		return