+++ ui/packages/shared/components/src/UserPreferences/index.tsx
@@ -15,32 +15,11 @@ import {cloneElement, useState} from 'react';
 
 import {Icon} from '@iconify/react';
 
-import {USER_PREFERENCES, useUIFeatureFlag} from '@parca/hooks';
+import {USER_PREFERENCES} from '@parca/hooks';
 
 import {Button, IconButton, Modal} from '../';
 import FlamegraphColorProfileSelector from './FlamegraphColorProfileSelector';
 import UserPreferenceItem from './UserPreferenceItem';
-
-interface FlagToggleProps {
-  name: string;
-  id: string;
-}
-
-const FlagToggle = ({name, id}: FlagToggleProps): JSX.Element => {
-  const [enabled, setEnabled] = useUIFeatureFlag(id);
-
-  return (
-    <div className="flex gap-2">
-      <input
-        type="checkbox"
-        id={id}
-        checked={enabled}
-        onChange={e => setEnabled(e.target.checked)}
-      />
-      <label htmlFor={id}>{name}</label>
-    </div>
-  );
-};
 
 const UserPreferences = ({customButton}: {customButton?: JSX.Element}): JSX.Element => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
@@ -79,10 +58,6 @@ const UserPreferences = ({customButton}: {customButton?: JSX.Element}): JSX.Elem
             userPreferenceDetails={USER_PREFERENCES.HIGHLIGHT_SIMILAR_STACKS}
           />
           <FlamegraphColorProfileSelector />
-          <div className="min-w-96 mt-10">
-            <h4 className="mb-2 font-medium">Experimental Features</h4>
-            <FlagToggle name="Enable Callgraph" id="callgraph" />
-          </div>
         </div>
         <div className="flex justify-end">
           <Button onClick={() => setIsOpen(false)} className="w-fit">