+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.735](https://github.com/parca-dev/parca/compare/@parca/web@0.16.734...@parca/web@0.16.735) (2024-07-30)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.734](https://github.com/parca-dev/parca/compare/@parca/web@0.16.733...@parca/web@0.16.734) (2024-07-30)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.733](https://github.com/parca-dev/parca/compare/@parca/web@0.16.732...@parca/web@0.16.733) (2024-07-29)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.732](https://github.com/parca-dev/parca/compare/@parca/web@0.16.731...@parca/web@0.16.732) (2024-07-29)
 
 **Note:** Version bump only for package @parca/web